+++ packages/services/src/Domain/Encryption/EncryptionProviderInterface.ts
@@ -20,7 +20,6 @@ import {
   KeySystemRootKeyParamsInterface,
   PortablePublicKeySet,
 } from '@standardnotes/models'
-import { PkcKeyPair } from '@standardnotes/sncrypto-common'
 
 export interface EncryptionProviderInterface {
   initialize(): Promise<void>
@@ -72,7 +71,6 @@ export interface EncryptionProviderInterface {
       }
   >
 
-  decryptErroredPayloads(): Promise<void>
   deleteWorkspaceSpecificKeyStateFromDevice(): Promise<void>
 
   unwrapRootKey(wrappingKey: RootKeyInterface): Promise<void>
@@ -109,9 +107,6 @@ export interface EncryptionProviderInterface {
     sharedVaultUuid: string | undefined,
     rootKeyToken: string,
   ): KeySystemItemsKeyInterface
-
-  getKeyPair(): PkcKeyPair
-  getSigningKeyPair(): PkcKeyPair
 
   asymmetricSignatureVerifyDetached(
     encryptedString: AsymmetricallyEncryptedString,