+++ packages/services/CHANGELOG.md
@@ -3,6 +3,10 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.63.23](https://github.com/standardnotes/app/compare/@standardnotes/services@1.63.22...@standardnotes/services@1.63.23) (2023-08-04)
+
+**Note:** Version bump only for package @standardnotes/services
+
 ## [1.63.22](https://github.com/standardnotes/app/compare/@standardnotes/services@1.63.21...@standardnotes/services@1.63.22) (2023-08-03)
 
 **Note:** Version bump only for package @standardnotes/services